backend/requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.2
flask-cors==6.0.2
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
psycopg==3.2.13
psycopg-binary==3.2.13
psycopg2-binary==2.9.11
python-dotenv==1.2.1
typing_extensions==4.15.0
Werkzeug==3.1.5
zipp==3.23.0
